Rename render.ts to gauge.ts
