+++ catalog/configurable-gauge/src/gauge.ts
